+++ classes/fbform.php
@@ -84,4 +84,17 @@ class fbform extends \mod_flexbook\form\base_form {
         $this->jump_section_fields(true);
         $this->close_form();
     }
+
+    /**
+     * Form validation.
+     *
+     * @param array $data
+     * @param array $files
+     * @return array
+     */
+    public function validation($data, $files) {
+        $errors = parent::validation($data, $files);
+        $errors = array_merge($errors, \local_ivpdfviewer\helper::validate_pdfviewer_elements($data, $files));
+        return $errors;
+    }
 }